extract common files
