+++ src/main/java/io/github/wmdietl/diagnostics/json/common/Range.java
@@ -1,7 +1,7 @@
-package io.github.wmdietl.diagnostics.json.lsp;
+package io.github.wmdietl.diagnostics.json.common;
 
 /**
- * LSP range in a text document expressed as (zero-based) start and end positions. A range is
+ * Range in a text document expressed as (zero-based) start and end positions. A range is
  * comparable to a selection in an editor. Therefore the end position is exclusive. If you want to
  * specify a range that contains a line including the line ending character(s) then use an end
  * position denoting the start of the next line.
